SOF-1747 - fix: switchActivateRundownDialog and switchRehearsalRundownDialog to DialogConfirmationService
