+++ src/app/rundown-view/components/rundown-header-context-menu/rundown-header-context-menu.component.ts
@@ -5,6 +5,7 @@ import { DialogService } from '../../../shared/services/dialog.service'
 import { ContextMenuOption } from '../../../shared/models/context-menu-option'
 import { DialogColorScheme, DialogSeverity } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component'
 import { RundownMode } from '../../../core/enums/rundown-mode'
+import { DialogConfirmationService } from '../../../shared/services/dialog-confirmation.service'
 
 @Component({
   selector: 'sofie-rundown-header-context-menu',
@@ -21,11 +22,11 @@ export class RundownHeaderContextMenuComponent {
   private readonly contextMenuOptionsForInactiveRundown: ContextMenuOption[] = [
     {
       label: 'Activate (On Air)',
-      contextAction: (): void => this.dialogService.switchActivateRundownDialog(this.rundown),
+      contextAction: (): void => this.dialogConfirmationService.switchActivateRundownDialog(this.rundown),
     },
     {
       label: 'Rehearse',
-      contextAction: (): void => this.dialogService.switchRehearsalRundownDialog(this.rundown),
+      contextAction: (): void => this.dialogConfirmationService.switchRehearsalRundownDialog(this.rundown),
     },
     {
       label: 'Reingest data',
@@ -55,7 +56,7 @@ export class RundownHeaderContextMenuComponent {
   private readonly contextMenuOptionsForRehearseRundown: ContextMenuOption[] = [
     {
       label: 'Activate (On Air)',
-      contextAction: (): void => this.dialogService.switchActivateRundownDialog(this.rundown),
+      contextAction: (): void => this.dialogConfirmationService.switchActivateRundownDialog(this.rundown),
     },
     ...this.contextMenuOptionsForActiveRundown,
   ]
@@ -73,7 +74,8 @@ export class RundownHeaderContextMenuComponent {
 
   constructor(
     private readonly rundownService: RundownService,
-    private readonly dialogService: DialogService
+    private readonly dialogService: DialogService,
+    private readonly dialogConfirmationService: DialogConfirmationService
   ) {}
 
   public openDeactivateRundownDialog(): void {
